refactor: move color scale to higher level component
